assembler directives (untested) + necessary changes to rest of program (all previous tests pass)

+++ expr_fns.c
@@ -1,10 +1,10 @@
-/* asm8085 (C) 2019 Marinus Oosters 
+/* asm8085 (C) 2019-20 Marinus Oosters 
  *
  * Implementations of functions and operators for expressions */
 
 #include "expr_fns.h"
 
-int eval_keyword(int kwdnum, int input) {
+intptr_t eval_keyword(int kwdnum, intptr_t input) {
     int k=0;
     #define _KWD(kwd) if(k==kwdnum) return evalkwd_##kwd (input); else k++;
     #include "operators.h"
@@ -12,7 +12,7 @@ int eval_keyword(int kwdnum, int input) {
     FATAL_ERROR("keyword does not exist #%d",kwdnum);
 }
 
-int eval_operator(int opnum, int *inputs) {
+intptr_t eval_operator(int opnum, intptr_t *inputs) {
     int o=0;
     #define _OPR(op, name, prcd, argsn) if(o==opnum) return evalop_##name (inputs); else o++;
     #include "operators.h"
@@ -21,13 +21,13 @@ int eval_operator(int opnum, int *inputs) {
 }
 
 // keywords
-int evalkwd_high(int input) {return ( ((unsigned) input) & 0xFF00) >> 8;}
-int evalkwd_low(int input)  {return ( ((unsigned) input) & 0x00FF);}
+intptr_t evalkwd_high(intptr_t input) {return ( ((unsigned) input) & 0xFF00) >> 8;}
+intptr_t evalkwd_low(intptr_t input)  {return ( ((unsigned) input) & 0x00FF);}
 
 // operators 
 
-#define UNOP(name, op)  int evalop_##name (int *inputs) { return op *inputs; }
-#define BINOP(name, op) int evalop_##name (int *inputs) { return inputs[0] op inputs[1]; }
+#define UNOP(name, op)  intptr_t evalop_##name (intptr_t *inputs) { return op *inputs; }
+#define BINOP(name, op) intptr_t evalop_##name (intptr_t *inputs) { return inputs[0] op inputs[1]; }
 
 BINOP(NE, !=)
 UNOP(BOOL_NOT, !)
